+++ musicxcst_downloader/src/musicxcst_downloader/__init__.py
@@ -1,3 +1,3 @@
 """MusicXCST Downloader desktop app."""
 
-__version__ = "4.0.9"
+__version__ = "4.1.0"